Safari-609.1.11.1/LayoutTests/platform/mac-wk1/imported/w3c/web-platform-tests/fetch/api/abort/general-serviceworker.https-expected.txt

CONSOLE MESSAGE: line 14: Unhandled Promise Rejection: TypeError: undefined is not an object (evaluating 'navigator.serviceWorker.getRegistration')

FAIL General fetch abort tests in a service worker undefined is not an object (evaluating 'navigator.serviceWorker.getRegistration')
